requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.52
botocore==1.19.52
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-storages==1.11.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
jsonfield==3.1.0
msgpack==1.0.0
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.2
